Move 'LocalFeedLoader' (and "FeedStore' Collaborator to it's own file)
