+++ EssentialFeed/EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -1,0 +1,49 @@
+//
+//  LocalFeedLoader.swift
+//  EssentialFeed
+//
+//  Created by PVC on 01/08/23.
+//
+
+import Foundation
+
+public class LocalFeedLoader {
+    private let store: FeedStore
+    private let currentDate: () -> Date
+    
+    public init(store: FeedStore, currentDate: @escaping () -> Date) {
+        self.store = store
+        self.currentDate = currentDate
+    }
+    
+    public func save(_ items: [FeedItem], completion: @escaping (Error?) -> Void ) {
+        store.deleteCachedFeed { [weak self] error in
+            guard let self = self else { return }
+            
+            if let cacheDeletionError = error {
+                completion(error)
+            } else {
+                self.cache(items: items, with: completion)
+            }
+        }
+    }
+    
+    private func cache(items: [FeedItem], with completion: @escaping (Error?) -> Void) {
+        store.insert(items, timestamp: currentDate()) { [weak self] error in
+            guard self != nil else { return }
+
+            completion(error)
+        }
+
+    }
+}
+
+public protocol FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+
+    func deleteCachedFeed(completion: @escaping DeletionCompletion)
+    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
+
+
